requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
fastapi>=0.95.0
uvicorn>=0.22.0
tensorboard>=2.12.0
wandb>=0.15.0
pytest>=7.3.0
pyyaml>=6.0
scikit-learn>=1.2.0
requests>=2.28.0
tqdm>=4.65.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
